refactor(setting): 重构设置组件，调整配置表单结构
- 移除 Setting 组件中复杂表单逻辑，拆分并移至 CustomConfigForm 组件中管理
- CustomConfigForm 组件内部新增表单状态和提交处理，简化父组件状态管理
- 删除全局配置相关冗余代码，改为组件内部使用 useGlobalConfig 读取和设置数据
- 保留 configMode 状态以控制显示 CustomConfigForm 或 QuotaCard 组件
- 优化 api key 显示切换逻辑，移至 CustomConfigForm 内部管理
- 增加对表单提交异常的错误提示和成功提示
- 通过 useEffect 根据选择的适配器动态设置默认参数
- 统一样式和表单元素的结构，提升组件可维护性和可读性
